requirements.txt
#python==3.7
numpy == 1.16
tensorflow-gpu == 1.*
gym == 0.7.3